+++ az-reportal/src/main/java/azkaban/jobtype/ReportalAbstractRunner.java
@@ -59,7 +59,6 @@ public abstract class ReportalAbstractRunner {
 
   public void run() throws Exception {
     System.out.println("Reportal: Setting up environment");
-    System.out.println("Reportal: " + this.props.toString());
 
     // Check the properties file
     if (this.props == null) {
